refactor(browserHistory): moving to store.js
