+++ app/client/components/pages/transferAccountListPage.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import { connect } from "react-redux";
 import styled, { ThemeProvider } from "styled-components";
 
-import { browserHistory } from "../../app.jsx";
+import { browserHistory } from "../../createStore.js";
 import {
   PageWrapper,
   WrapperDiv,
