+++ app/client/createStore.js
@@ -1,6 +1,7 @@
 import { createStore, applyMiddleware, compose } from "redux";
 import createSagaMiddleware from "redux-saga";
 import { delayConfiguration } from "pusher-redux";
+import { createBrowserHistory } from "history";
 import * as Sentry from "@sentry/browser";
 import { version } from "../package.json";
 
@@ -8,6 +9,7 @@ import appReducer from "./reducers/rootReducer";
 import rootSaga from "./sagas/rootSaga";
 
 const sagaMiddleware = createSagaMiddleware();
+export const browserHistory = createBrowserHistory({});
 
 // Setup redux dev tools
 const composeSetup =
